refactor : 함수명 변경과 중복 사용되는 상수 분리
